+++ src/smirks/reference.js
@@ -39,7 +39,7 @@ export const reactionTemplates = {
   },
   imineReduction: {
     name: 'Imine Reduction',
-    smirks: '[C;!$([C](=[N])[N,O,S]):1]=[N;!$([N]-[N,O,S]):2]>>[CH2:1][NH:2]'
+    smirks: '[C;!$([C](=[N])[N,O,S]):1]=[N;!$([N]-[N,O,S]):2]>>[C:1][N:2]'
   },
   alkeneHydrogenation: {
     name: 'Alkene Hydrogenation',
@@ -81,11 +81,11 @@ export const reactionTemplates = {
   amideHydrolysis: { name: 'Amide Hydrolysis', smirks: '[C:1](=[O:2])[N:3]>>[C:1](=[O:2])O.[N:3]' },
   amineAcylation: {
     name: 'Amine Acylation',
-    smirks: '[C:1](=[O:2])[Cl:3].[NH2+0;!$([N]-[C](=O)):4]>>[C:1](=[O:2])[NH+0:4].[ClH0-:3]'
+    smirks: '[C:1](=[O:2])[Cl:3].[N+0;!H0;!$([N]-[C](=O)):4]>>[C:1](=[O:2])[N+0:4].[ClH0-:3]'
   },
   amineAlkylation: {
     name: 'Amine Alkylation',
-    smirks: '[C;X4;H2,H3:1][Cl:2].[NH2+0;!$([N]-[C](=O)):3]>>[C:1][NH+0:3].[ClH0-:2]'
+    smirks: '[C;X4;H2,H3:1][Cl:2].[N+0;!H0;!$([N]-[C](=O)):3]>>[C:1][N+0:3].[ClH0-:2]'
   },
   imineHydrolysis: { name: 'Imine Hydrolysis', smirks: '[C:1]=[N:2]>>[C:1]=O.[N:2]' },
   nitrileHydrolysisToAmide: { name: 'Nitrile Hydrolysis To Amide', smirks: '[C:1]#[N:2]>>[C:1](=O)[N:2]' },
@@ -116,9 +116,9 @@ export const reactionTemplates = {
     name: 'Sulfoxide Oxidation To Sulfone',
     smirks: '[C:2][S:1](=[O:4])[C:3]>>[C:2][S:1](=[O:4])(=O)[C:3]'
   },
-  amineProtonation: { name: 'Amine Protonation', smirks: '[NH2+0;!$([N]-[C](=O)):1]>>[NH3+:1]' },
+  amineProtonation: { name: 'Amine Protonation', smirks: '[N+0;!$([N]-[C](=O)):1]>>[N+:1]' },
   ammoniumDeprotonation: { name: 'Ammonium Deprotonation', smirks: '[N+;!H0:1]>>[N+0:1]' },
   phenolDeprotonation: { name: 'Phenol Deprotonation', smirks: '[c:1][OH:2]>>[c:1][OH0-:2]' },
   phenolateProtonation: { name: 'Phenolate Protonation', smirks: '[c:1][OH0-:2]>>[c:1][OH+0:2]' },
-  nitroReduction: { name: 'Nitro Reduction', smirks: '[N+:1](=[O:2])[O-:3]>>[NH2+0:1]' }
+  nitroReduction: { name: 'Nitro Reduction', smirks: '[N+:1](=[O:2])[O-:3]>>[N+0:1].[OH2+0:2].[OH2+0:3]' }
 };